chore: Rename file & fix up routing configs
- Removed the `mostPlayed` translations (we need to migrate the other languages over).
