+++ mobile/src/modules/insights/screens/RecapView.tsx
